+++ http/src/main/scala/com/twitter/finatra/http/exceptions/ExceptionMapperCollection.scala
@@ -4,9 +4,9 @@ import scala.collection.mutable.ArrayBuffer
 
 /**
  * Represents a collection of [[com.twitter.finatra.http.exceptions.ExceptionMapper]]s
- * which is a {{{Traversable[Manifest[ExceptionMapper[_]]]}}}.
+ * which is an {{{Iterable[Manifest[ExceptionMapper[_]]]}}}.
  */
-class ExceptionMapperCollection extends Traversable[Manifest[ExceptionMapper[_]]] {
+class ExceptionMapperCollection extends Iterable[Manifest[ExceptionMapper[_]]] {
 
   private[this] val manifests = ArrayBuffer[Manifest[ExceptionMapper[_]]]()
 
@@ -21,5 +21,5 @@ class ExceptionMapperCollection extends Traversable[Manifest[ExceptionMapper[_]]
 
   override def foreach[U](f: (Manifest[ExceptionMapper[_]]) => U): Unit = manifests.foreach(f)
 
-  override def seq: Traversable[Manifest[ExceptionMapper[_]]] = manifests.seq
+  override def iterator: Iterator[Manifest[ExceptionMapper[_]]] = manifests.iterator
 }